Consolidate hosting code
We made a copy of the hosting code for the UCP repo, so we now have two
copies in the Radius repo.

Ultimately UCP will act as a library for the RPs to build-upon when we
split the repositories again - so the hosting code should live under
UCP.
