+++ cmd/appcore-async/main.go
@@ -17,8 +17,8 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/project-radius/radius/pkg/corerp/backend"
 	"github.com/project-radius/radius/pkg/corerp/hostoptions"
-	"github.com/project-radius/radius/pkg/hosting"
 	"github.com/project-radius/radius/pkg/radlogger"
+	"github.com/project-radius/radius/pkg/ucp/hosting"
 )
 
 func main() {
